test(editor): move the Editor tests into src/lib/editor/test
