+++ packages/editor/src/lib/editor/test/Editor.assets.test.ts
@@ -13,8 +13,8 @@ import {
 	atom,
 	createTLCurrentUser,
 	createTLStore,
-} from '../..'
-import { Editor } from './Editor'
+} from '../../..'
+import { Editor } from '../Editor'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
