+++ packages/editor/src/lib/editor/test/Editor.deepLinks.test.ts
@@ -9,8 +9,8 @@ import {
 	T,
 	TLShape,
 	createShapeId,
-} from '../..'
-import { TestEditor } from '../test/TestEditor'
+} from '../../..'
+import { TestEditor } from '../../test/TestEditor'
 
 const BOX_TYPE = 'my-custom-shape'
