+++ packages/editor/src/lib/editor/test/Editor.following.test.ts
@@ -18,8 +18,8 @@ import {
 	createTLCurrentUser,
 	createTLStore,
 	createUserId,
-} from '../..'
-import { Editor, TLEditorOptions } from './Editor'
+} from '../../..'
+import { Editor, TLEditorOptions } from '../Editor'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
