+++ packages/editor/src/lib/editor/test/Editor.history.test.ts
@@ -11,8 +11,8 @@ import {
 	createShapeId,
 	createTLCurrentUser,
 	createTLStore,
-} from '../..'
-import { Editor } from './Editor'
+} from '../../..'
+import { Editor } from '../Editor'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
