+++ packages/editor/src/lib/editor/test/Editor.selection.test.ts
@@ -13,8 +13,8 @@ import {
 	createShapeId,
 	createTLCurrentUser,
 	createTLStore,
-} from '../..'
-import { Editor } from './Editor'
+} from '../../..'
+import { Editor } from '../Editor'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
