+++ packages/editor/src/lib/editor/test/Editor.test.ts
@@ -9,9 +9,9 @@ import {
 	TLShape,
 	createShapeId,
 	createTLStore,
-} from '../..'
-import { Editor } from './Editor'
-import { StateNode } from './tools/StateNode'
+} from '../../..'
+import { Editor } from '../Editor'
+import { StateNode } from '../tools/StateNode'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
