+++ packages/editor/src/lib/editor/test/Editor.themes.test.ts
@@ -15,8 +15,8 @@ import {
 	createTLCurrentUser,
 	createTLStore,
 	react,
-} from '../..'
-import { Editor, TLEditorOptions } from './Editor'
+} from '../../..'
+import { Editor, TLEditorOptions } from '../Editor'
 
 const MY_CUSTOM_SHAPE_TYPE = 'my-custom-shape'
